Clarify glPortal's development: sporadic
The glPortal team is pretty small and due to wrong design decisions
(which are being corrected) the project has failed to attract more
people. As a result, development is slow.

Also removes the "status: playable" info since the last actually
playable version is very old and does not reflect at all the recent work
that has been done on the game and its engine.

In fact, as of this commit, portals (the central game element) have been
broken for months.