Update OpenRCT2 Entry (#510)
* Update OpenRCT2 Entry

I updated the OpenRCT2 Entry to reflect the current status of the project. Is there a way, without copying and pasting of course, that I can indicate that the game is capable of replacing RollerCoaster Tycoon 1 as well? (Importing and running Scenarios/Saves/etc)

Additionally, we are technically a clone of "RollerCoaster Tycoon Classic" on iOS, Android, and Windows due to it being a re-release of the original game with a new UI on those platforms. We are also able to import this games assets as a base.

* Removed very old screenshot.