Fix image and video gallery when sorted by updated date
Changed tr an event delegation strategy to handle the toggle of image
and video galleries.

When you clone a node in JavaScript, it doesn't copy attached events to
the new node.

Instead of attaching a new event for each game, It's attached to the
parent node to handle the click event when the css-class matches "toggle".

fix #1193